Implemented Templates in Django
the default.html file within templates/layout contains all of the site
wide code. The individual pages in templates contain something called
blocks that act as a place holder for content. So for example index.html
has a pagetitle block and a primary block, the content in these blocks
get inserted into the place holder blocks within default when the page
is rendered. So for each additional page we add you just need to extend
the default.html and create blocks to make the template work. Extend
Syntax: {% extends 'layout\default.html' %}